refactor(sdk): move the common pagination types to the paginators root mod
