+++ crates/matrix-sdk/src/paginators/mod.rs
@@ -17,4 +17,72 @@
 mod room;
 pub mod thread;
 
+use matrix_sdk_base::deserialized_responses::TimelineEvent;
 pub use room::*;
+use ruma::OwnedEventId;
+
+/// Pagination token data, indicating in which state is the current pagination.
+#[derive(Clone, Debug, PartialEq)]
+pub enum PaginationToken {
+    /// We never had a pagination token, so we'll start back-paginating from the
+    /// end, or forward-paginating from the start.
+    None,
+    /// We paginated once before, and we received a prev/next batch token that
+    /// we may reuse for the next query.
+    HasMore(String),
+    /// We've hit one end of the timeline (either the start or the actual end),
+    /// so there's no need to continue paginating.
+    HitEnd,
+}
+
+impl From<Option<String>> for PaginationToken {
+    fn from(token: Option<String>) -> Self {
+        match token {
+            Some(val) => Self::HasMore(val),
+            None => Self::None,
+        }
+    }
+}
+
+/// The result of a single event pagination.
+#[derive(Debug)]
+pub struct PaginationResult {
+    /// Events returned during this pagination.
+    ///
+    /// If this is the result of a backward pagination, then the events are in
+    /// reverse topological order.
+    ///
+    /// If this is the result of a forward pagination, then the events are in
+    /// topological order.
+    pub events: Vec<TimelineEvent>,
+
+    /// Did we hit *an* end of the timeline?
+    ///
+    /// If this is the result of a backward pagination, this means we hit the
+    /// *start* of the timeline.
+    ///
+    /// If this is the result of a forward pagination, this means we hit the
+    /// *end* of the timeline.
+    pub hit_end_of_timeline: bool,
+}
+
+/// An error that happened when using a [`Paginator`].
+#[derive(Debug, thiserror::Error)]
+pub enum PaginatorError {
+    /// The target event could not be found.
+    #[error("target event with id {0} could not be found")]
+    EventNotFound(OwnedEventId),
+
+    /// We're trying to manipulate the paginator in the wrong state.
+    #[error("expected paginator state {expected:?}, observed {actual:?}")]
+    InvalidPreviousState {
+        /// The state we were expecting to see.
+        expected: PaginatorState,
+        /// The actual state when doing the check.
+        actual: PaginatorState,
+    },
+
+    /// There was another SDK error while paginating.
+    #[error("an error happened while paginating: {0}")]
+    SdkError(#[from] Box<crate::Error>),
+}
